Server list implemented
